backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic[email]
python-multipart
passlib[bcrypt]
python-jose[cryptography]
aiomysql
pymysql